Add files via upload
The project focuses on creating a scaled-down 3D topographic model of the Trishuli Hydropower Project area. Using contour lines from topographic maps, the students built a physical model to demonstrate the region's elevation and major components like dam, powerhouse, penstock, etc. The scale used was 1:2000 for elevation and 1:5700 for length